First cut at compressing all of Value's memory into 8 bytes

+++ cvs/objects/util/base/include/manage_state_variables.hpp
@@ -110,7 +110,7 @@ private:
     
     //! The total number of individual Values flagged as STATE that could possibly
     //! be changed during World.calc( mPeriodToCollect ).
-    size_t mNumCollected;
+    uint64_t mNumCollected;
     
     //! The list of individual Values flagged as STATE that could possibly be
     //! changed during World.calc( mPeriodToCollect ).  We store them in a list
